Return formulation after solving it
Update the ModelFactory class to return the formulation after it has
been solved. Drop the ALL entry in the Solver enum and create the
solveInAllSolvers method to provide the same functionality. If using
this method, it will return a list of formulations solved in the all
solvers, while using solveIn will only return a single instance of the
formulation.